fix Linux tests
