+++ Packages/CatbirdAPI/Sources/CatbirdAPI/CatbirdError.swift
@@ -1,7 +1,7 @@
 import Foundation
 
 /// From `Vapor.ErrorMiddleware`.
-private struct ErrorResponse: Codable {
+struct ErrorResponse: Codable {
 
     /// Always `true` to indicate this is a non-typical JSON response.
     let error: Bool
@@ -20,16 +20,6 @@ public struct CatbirdError: LocalizedError, CustomNSError {
 
     /// A localized message describing the reason for the failure.
     public let failureReason: String?
-
-#if !os(Linux)
-    init?(response: HTTPURLResponse, data: Data?) {
-        guard !(200..<300).contains(response.statusCode) else { return nil }
-        self.errorCode = response.statusCode
-        self.failureReason = data.flatMap { (body: Data) in
-            try? JSONDecoder().decode(ErrorResponse.self, from: body).reason
-        }
-    }
-#endif
 
     /// A localized message describing what error occurred.
     public var errorDescription: String? {
