--- Packages/CatbirdAPI/Sources/CatbirdAPI/URLSessionTask+Wait.swift
@@ -1,18 +1,0 @@
-#if !os(Linux)
-
-import Foundation
-
-extension URLSessionTask {
-
-    /// Wait until task completed.
-    func wait() {
-        guard let timeout = currentRequest?.timeoutInterval else { return }
-        let limitDate = Date(timeInterval: timeout, since: Date())
-        while state == .running && RunLoop.current.run(mode: .default, before: limitDate) {
-            // wait
-        }
-    }
-
-}
-
-#endif
